requirements.txt
Django==4.2.10
djangorestframework==3.14.0
apscheduler==3.10.4
django-apscheduler==0.6.2
psycopg2-binary==2.9.10
gunicorn==20.0.4
requests==2.28.1
whitenoise==6.5.0
django-cors-headers==4.7.0